Removed click templates from data
These are no longer necessary because I discovered functions in PAMpal that will allow me to query the studies by UID to plot waveforms & spectrograms of individual clicks. That was the whole point of collecting the binary data into this data object. So that data object is now extraneous, and I have deleted it.